requirements/base.txt
Django<=3.3

psycopg2-binary==2.8.6
requests==2.25.1
django-model-utils==4.1.1
django-markdown-deux==1.0.5
django-localflavor==3.1
django-extensions==3.1.3
djangorestframework==3.12.4
django-redis==5.0.0
freezegun==1.1.0
hiredis==2.0.0
redis==3.5.3
Pillow==8.3.1
icalendar==4.0.7
dealer==2.1.0
django-robots==4.0
django-debug-toolbar==3.2.1
newspaper3k==0.2.8

https://github.com/DemocracyClub/dc_django_utils/archive/refs/tags/0.1.4.tar.gz
https://github.com/DemocracyClub/dc_signup_form/archive/refs/tags/2.1.1.tar.gz
https://github.com/DemocracyClub/design-system/archive/refs/tags/0.1.6.tar.gz
https://github.com/DemocracyClub/dc_django_utils/archive/refs/tags/0.1.4.tar.gz

djangorestframework-jsonp==1.0.2
feedparser==6.0.8

sentry-sdk==1.3.1
uk-election-timetables==2.2.2
aws-sam-cli==1.27.2
awscli==1.20.14
django-dotenv==1.4.2
